saved_restaurants_dashboard (#71): Creating a diagram and markdown file describing the workflow of the Saved Restaurants Dashboard.
This markdown file describes the work flow from grabbing the data, to displaying it in the Want To Try and Visited sections, how it moves restaurants in between sections, how it removes restaurants and how it links to IndexedDB and the Review System. The mermaid syntax graph I created shows clearly how they all relate for all these different actions based on user's choice.