ranking_bracket_system (#15, #16, #19, #20, #81, #82, #83) : updated my branch with ranking_bracket_system js and css
My feature is the ranking bracket system. The goal of this feature is to narrow down the restaurant options until the most ideal restaurant for the user is chosen. There will be eight restaurant options that will be rotated through the voting options. There are save and vote buttons under each restaurant option. When the user clicks on the vote button, the restaurant that they voted for stays and the other restaurant changes to another option. This continues until there are no more restaurants. The last restaurant that was voted for is the “winner” and it will show the user that it is the best restaurant for them. When the save button is clicked, that restaurant that it was clicked for will be saved to a list of saved restaurants. There is also a “show saved restaurants” button. When this button is clicked, all the previously saved restaurants will show up. Also, a “hide saved restaurants” button will appear, and when this button is clicked, the saved restaurants will be hidden. Furthermore, will the saved restaurants are shown, if you click on a restaurant, there will be an option to delete that restaurant from the saved list. If the user chooses to delete the restaurant, then it will be removed from the saved restaurants.