ranking_bracket_system (#91): accessing the restaurants from localStorage instead of hardcoding it.
Instead of hardcoding restaurants to use in the ranking bracket system, we access the data in localStorage to use  as voting options.  It is going to access the restaurants that the user voted yes on from the feature recommendation system. Each restaurant that is voted "yes" on will be used in this voting system. The restaurant that is voted on will stay and the other restaurant will change. The last restaurant that is voted on will be the "ideal" restaurant for the user.