m4_rating_system (#115): populate prefilled data on rating page
On dashboard.html, when the user clicks the arrow to review a restaurant, changed the text box allowing the user to type in the restaurant to an already populated box of the restaurant chosen. Also, kept the review there so that the user can find their past review instead of clearing the data after submission.