Authorization #132, #133: implemented register and login feature so that the user can register an account and log into its account
The user should be able to enter any username and password in order to register to Plateful. If the username is already in use, a message will pop up and show that the username is already taken. The user should then be able to log in to their account using the username and password they registered with. If they enter the wrong username/password, a message will pop up letting them know that one of them is invalid.